python/requirements.txt
sentence-transformers==3.3.1
faiss-cpu==1.9.0
requests==2.32.3
numpy==1.26.4
matplotlib==3.9.2
tabulate==0.9.0
PyPDF2==3.0.1
colorama==0.4.6
